+++ migrations/1_Coin.js
@@ -1,4 +1,4 @@
-const Coin = artifacts.require('ERC20BaseMock');
+const Coin = artifacts.require('ERC20FullMock');
 
 module.exports = async (deployer, network, accounts) => {
     const initialBalance = "1000000";