+++ CHANGELOG.md
@@ -5,10 +5,15 @@
 ### Breaking changes
  * Migration to `@animoca/ethereum-contracts-core_library:1.0.0` with `solc:0.6.x` and `@openzeppelin/contracts:3.x`.
  * Moved `WhitelistedOperators.sol` to `@animoca/ethereum-contracts-core_library`.
+ * Renamed `ERC20Base` to `ERC20Full`.
 
 ### New features
+ * Added openezeppelin-based `ERC20.sol` implementation.
  * Added `IERC20Detailed.sol`.
  * Added example migrations.
+
+### Improvements
+ * `ERC20Fees` now inherits `PayoutWallet` contract.
 
 ## 0.0.2 (14/04/2020)